make segment pretty

+++ project_folder/scripts/segment.py
@@ -1,10 +1,10 @@
-from pathlib import Path
-from spacy.cli._util import Arg
-import typer
-from rich.progress import track
-from PIL import Image
-
 import torch
+import typer
+from PIL import Image
+from pathlib import Path
+from rich.progress import track
+from spacy.cli._util import Arg
+
 from kraken import blla
 from kraken import serialization
 
@@ -19,12 +19,11 @@ def segment(
 
     images = list(collection_path.glob("**/*.jpg"))
     for image in track(images, description="Segmenting images..."):
-        filename = image.name
-        img = Image.open(image)
-        # segment the image using kraken
-        baseline_seg = blla.segment(img, device=device)
-        alto_xml = serialization.serialize_segmentation(baseline_seg, image_name=filename, image_size=img.size, template='alto')
-        image.with_suffix('.xml').write_text(alto_xml)
+        if not image.with_suffix('.xml').exists():
+            img = Image.open(image)
+            baseline_seg = blla.segment(img, device=device)
+            alto_xml = serialization.serialize_segmentation(baseline_seg, image_name=image.name, image_size=img.size, template='alto')
+            image.with_suffix('.xml').write_text(alto_xml)
 
 if __name__ == "__main__":
     typer.run(segment)
